Create some tests that run on RBE with platforms with various exec_properties that are populated using create_exec_properties_dict()
Currently, I added only tests for docker_privileged=True/False and for docker_network_enabled=True/False. But this PR proivides a skeleton that demonstrates how other properties should be tested.

In order to run these tests, bazel needs to run with
--extra_execution_platforms=@rbe_default//config:platform,//tests/rules/experimental/rbe:platform_privileged_on,//tests/rules/experimental/rbe:platform_network_enabled

I have not yet piped these test into any pipeline.
One option is to pipe these tests as presubmit tests of bazel-toolchains. This cannot yet be done because these tests can only run on bazel 0.29 and up and that container does not yet exist. The reason why it is not compatible with older bazel versions is because platform does not have an exec_properties in those versions.

However, what I really want to do is to pipe these tests as bazel buildkite downstream tests (see https://github.com/bazelbuild/continuous-integration/blob/master/buildkite/README.md#configuring-a-pipeline).
In order to do that, .bazelci/presubmit.yml needs to first be converted to a newer format. I plan to do this in a separate pull request.