Rename the docker_network_enabled param to docker_enabled.
Instead of having docker_network_enabled=True/False, have instead docker_enabled=standard/off.
The reason for this change is to be more aligned with the RBE key/value name.
I want to keep this layer of abstraction as close as possible to the RBE layer so as to avoid the need to maintain extra documentation regarding what the paraters and their values are, but instead just link to the RBE documentation.
This is done in the context of a proposal to add an extra layer of abstraction above create_exec_properties_dict, which will abstract away the RBE implementation and could be extended to support other remote backends.

The layer of abstraction provided by create_exec_properties_dict will now only:
1. Wrap the RBE key names with the more bzl standard snake_case key names.
2. Provide some sanity checks for the keys and the values.

But will no longer:
3. Attempt to simplify or change the key name or the values.