Fix broken links and fix docs. (#752)
The broken links should be fixed as a result of the RBE execution properties having moved to a separate page.

The fix in the doc is a result of me realizing that sometimes there is a valid reason to call create_exec_properties_dict from a BUILD file. Specifically, this is fine if it is done while creating a platform.
The crucial distinction here is that it should not be done in a target because this target might be depended on my a different repo. A platform can also be depended on, but will typically be more easily overridable.