Remove dependency from rules/rbe_repo/build_gen.bzl to skylib
Despite the fact that we are still deprecating merge_dicts and recommending that users use dict.add from skylib, we decided not to add that dependency from rules/rbe_repo/build_gen.bzl and instead use a locally defined _merge_dicts.

The reason for doing this is that it complicated the setup for any WORKSAPCE that only want to use rbe_autconfig. We don't want these simple WORKSPACEs to need to call respositories() in order not to break.