Replace create_exec_properties_dict with create_rbe_exec_properties_dict (#763)
* Replace create_exec_properties_dict with create_rbe_exec_properties_dict

And also replace merge_dicts with dicts.add

* Change initialization order.

In repositories/repositories.bzl in the function repositories(), I reversed the order of handling io_bazel_rules_docker" and "bazel_skylib"

* Add bazel_toolchains_repositories to bazel_toolchains_client WORKSPACE

* Remove bazel symlinks that were unintentionally introduced in prev commit.

Also add them to .gitignore.

* Modify gitignore

* Add repositories to tests/rbe_repo/examples/toolchain_config_host/WORKSPACE as well.

* Remove dependency from rules/rbe_repo/build_gen.bzl to skylib

Despite the fact that we are still deprecating merge_dicts and recommending that users use dict.add from skylib, we decided not to add that dependency from rules/rbe_repo/build_gen.bzl and instead use a locally defined _merge_dicts.

The reason for doing this is that it complicated the setup for any WORKSAPCE that only want to use rbe_autconfig. We don't want these simple WORKSPACEs to need to call respositories() in order not to break.