Optimize RBE Container image existence check (#892)
* Optimize RBE Container image existence check

Use a custom Go binary to use the image manifest to verify image
existence instead of using docker pull to download all the images. The
list of image monotonically increases and the test is currently timing
out because it takes more than 30mins. The new implementation takes 11s
and gives us more room for growth.

* buildifier

* buildifier